flask-server/requirements.txt
langchain
langchain_groq 
langchain_core 
langchain_community
paddleocr 
paddlepaddle
requests 
pillow